+++ src/main/java/net/brnbrd/delightful/common/item/knife/Knives.java
@@ -1,36 +1,48 @@
 package net.brnbrd.delightful.common.item.knife;
 
-import net.brnbrd.delightful.Util;
 import net.brnbrd.delightful.common.item.DelightfulItems;
 import net.brnbrd.delightful.common.item.DelightfulTiers;
-import net.brnbrd.delightful.common.item.knife.allthemodium.AllthemodiumKnifeItem;
-import net.brnbrd.delightful.common.item.knife.create_sa.BlazingKnifeItem;
-import net.brnbrd.delightful.common.item.knife.create_sa.ExperienceKnifeItem;
-import net.brnbrd.delightful.common.item.knife.create_sa.GildedQuartzKnifeItem;
-import net.brnbrd.delightful.common.item.knife.enlightened_end.AdamantiteKnifeItem;
-import net.brnbrd.delightful.common.item.knife.forbidden_arcanus.DracoArcanusKnifeItem;
-import net.brnbrd.delightful.common.item.knife.lolenderite.ObsdianInfusedEnderiteKnifeItem;
-import net.brnbrd.delightful.common.item.knife.nethers_exoticism.KiwanoKnifeItem;
-import net.brnbrd.delightful.common.item.knife.oresabovediamonds.LargeAmethystKnifeItem;
-import net.brnbrd.delightful.common.item.knife.phantasm.StelliumKnifeItem;
-import net.brnbrd.delightful.common.item.knife.rootsclassic.LivingKnifeItem;
-import net.brnbrd.delightful.common.item.knife.seeds.LeafKnifeItem;
-import net.brnbrd.delightful.common.item.knife.twilightforest.FieryKnifeItem;
-import net.brnbrd.delightful.common.item.knife.twilightforest.IronwoodKnifeItem;
-import net.brnbrd.delightful.common.item.knife.twilightforest.SteeleafKnifeItem;
+import net.brnbrd.delightful.common.item.knife.compat.additionaladditions.GildedNetheriteKnifeItem;
+import net.brnbrd.delightful.common.item.knife.compat.additionaladditions.RoseGoldKnifeItem;
+import net.brnbrd.delightful.common.item.knife.compat.allthemodium.AllthemodiumKnifeItem;
+import net.brnbrd.delightful.common.item.knife.compat.botania.ElementiumKnifeItem;
+import net.brnbrd.delightful.common.item.knife.compat.botania.ManasteelKnifeItem;
+import net.brnbrd.delightful.common.item.knife.compat.botania.TerraKnifeItem;
+import net.brnbrd.delightful.common.item.knife.compat.byg.PendoriteKnifeItem;
+import net.brnbrd.delightful.common.item.knife.compat.create_sa.BlazingKnifeItem;
+import net.brnbrd.delightful.common.item.knife.compat.create_sa.ExperienceKnifeItem;
+import net.brnbrd.delightful.common.item.knife.compat.create_sa.GildedQuartzKnifeItem;
+import net.brnbrd.delightful.common.item.knife.compat.deeperdarker.WardenKnifeItem;
+import net.brnbrd.delightful.common.item.knife.compat.enlightened_end.AdamantiteKnifeItem;
+import net.brnbrd.delightful.common.item.knife.compat.forbidden_arcanus.DracoArcanusKnifeItem;
+import net.brnbrd.delightful.common.item.knife.compat.forbidden_arcanus.ReinforcedDeorumKnifeItem;
+import net.brnbrd.delightful.common.item.knife.compat.lolenderite.EnderiteKnifeItem;
+import net.brnbrd.delightful.common.item.knife.compat.lolenderite.ObsdianInfusedEnderiteKnifeItem;
+import net.brnbrd.delightful.common.item.knife.compat.mythicbotany.AlfKnifeItem;
+import net.brnbrd.delightful.common.item.knife.compat.nethers_exoticism.KiwanoKnifeItem;
+import net.brnbrd.delightful.common.item.knife.compat.nourished_nether.NecroniumKnifeItem;
+import net.brnbrd.delightful.common.item.knife.compat.oresabovediamonds.LargeAmethystKnifeItem;
+import net.brnbrd.delightful.common.item.knife.compat.oresabovediamonds.NetheriteOpalKnifeItem;
+import net.brnbrd.delightful.common.item.knife.compat.phantasm.StelliumKnifeItem;
+import net.brnbrd.delightful.common.item.knife.compat.rootsclassic.LivingKnifeItem;
+import net.brnbrd.delightful.common.item.knife.compat.seeds.LeafKnifeItem;
+import net.brnbrd.delightful.common.item.knife.compat.spirit.SoulSteelKnifeItem;
+import net.brnbrd.delightful.common.item.knife.compat.twilightforest.FieryKnifeItem;
+import net.brnbrd.delightful.common.item.knife.compat.twilightforest.IronwoodKnifeItem;
+import net.brnbrd.delightful.common.item.knife.compat.twilightforest.KnightmetalKnifeItem;
+import net.brnbrd.delightful.common.item.knife.compat.twilightforest.SteeleafKnifeItem;
+import net.brnbrd.delightful.common.item.knife.compat.undergarden.ForgottenKnifeItem;
+import net.brnbrd.delightful.common.item.knife.compat.undergarden.FroststeelKnifeItem;
+import net.brnbrd.delightful.common.item.knife.compat.undergarden.UtheriumKnifeItem;
+import net.brnbrd.delightful.compat.BotaniaCompat;
 import net.brnbrd.delightful.compat.Mods;
 import net.brnbrd.delightful.data.tags.DelightfulItemTags;
 import net.minecraft.ChatFormatting;
-import net.minecraft.network.chat.Component;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.item.Item;
-import net.minecraft.world.item.crafting.Ingredient;
 import net.minecraftforge.common.Tags;
 import net.minecraftforge.registries.RegistryObject;
 import vectorwing.farmersdelight.FarmersDelight;
-import vectorwing.farmersdelight.common.registry.ModItems;
-
-import java.util.function.Supplier;
 
 public class Knives extends DelightfulItems {
 
@@ -54,69 +66,92 @@ public class Knives extends DelightfulItems {
 	public static final RegistryObject<Item> NICKEL = registerIngotKnife("nickel");
 	public static final RegistryObject<Item> ZINC = registerIngotKnife("zinc");
 
-	public static final RegistryObject<Item> MYTHRIL = registerCompatKnife("mythril", Mods.SO, ingot("mythril"));
-	public static final RegistryObject<Item> ADAMANTIUM = registerCompatKnife("adamantium", Mods.SO, ingot("adamantium"));
+	public static final RegistryObject<Item> MYTHRIL = registerCompatIngotKnife("mythril", Mods.SO);
+	public static final RegistryObject<Item> ADAMANTIUM = registerCompatIngotKnife("adamantium", Mods.SO);
 	public static final RegistryObject<Item> ONYX = registerCompatKnife("onyx", Mods.SO, gem("onyx"));
-	public static final RegistryObject<Item> THYRIUM = registerCompatKnife("thyrium", Mods.FUS, ingot("thyrium"));
-	public static final RegistryObject<Item> SINISITE = registerCompatKnife("sinisite", Mods.FUS, ingot("sinisite"));
-	public static final RegistryObject<Item> ALLTHEMODIUM = registerItem("allthemodium_knife", () -> new AllthemodiumKnifeItem((new Item.Properties()).tab(FarmersDelight.CREATIVE_TAB)));
-	public static final RegistryObject<Item> ENDERITE = registerSmithedKnife("enderite", Util.ing(ModItems.NETHERITE_KNIFE), ingot("enderite"));
-	public static final RegistryObject<Item> DEORUM = registerCompatKnife("deorum", Mods.FA, ingot("deorum"));
-	public static final RegistryObject<Item> REINFORCED_DEORUM = registerSmithedKnife("reinforced_deorum", Util.ing(DEORUM), DelightfulItemTags.STELLARITE_PIECE, "forbidden_arcanus");
-	public static final RegistryObject<Item> DRACO_ARCANUS = registerItem("draco_arcanus_knife", () -> new DracoArcanusKnifeItem((new Item.Properties()).tab(FarmersDelight.CREATIVE_TAB)));
-	public static final RegistryObject<Item> OSMIUM = registerCompatKnife("osmium", Mods.MEKT, ingot("osmium"));
-	public static final RegistryObject<Item> REFINED_GLOWSTONE = registerCompatKnife("refined_glowstone", Mods.MEKT, ingot("refined_glowstone"));
-	public static final RegistryObject<Item> REFINED_OBSIDIAN = registerCompatKnife("refined_obsidian", Mods.MEKT, ingot("refined_obsidian"));
-	public static final RegistryObject<Item> OBSIDIAN_INFUSED_ENDERITE = registerItem("obsidian_infused_enderite_knife", () -> new ObsdianInfusedEnderiteKnifeItem((new Item.Properties()).tab(FarmersDelight.CREATIVE_TAB)));
-	public static final RegistryObject<Item> NETHERITE_OPAL = registerSmithedKnife("netherite_opal", Util.ing(BLACK_OPAL), Tags.Items.INGOTS_NETHERITE, "oresabovediamonds");
-	public static final RegistryObject<Item> LARGE_AMETHYST = registerItem("large_amethyst_knife", () -> new LargeAmethystKnifeItem((new Item.Properties()).tab(FarmersDelight.CREATIVE_TAB)));
-	public static final RegistryObject<Item> FIERY = registerItem("fiery_knife", () -> new FieryKnifeItem((new Item.Properties()).tab(FarmersDelight.CREATIVE_TAB)));
-	public static final RegistryObject<Item> IRONWOOD = registerItem("ironwood_knife", () -> new IronwoodKnifeItem((new Item.Properties()).tab(FarmersDelight.CREATIVE_TAB)));
-	public static final RegistryObject<Item> KNIGHTMETAL = registerCompatKnife("knightmetal", Mods.TF, ingot("knightmetal"), Component.translatable("item." + Mods.TF + ".knightmetal_sword.tooltip").withStyle(ChatFormatting.GRAY));
-	public static final RegistryObject<Item> STEELEAF = registerItem("steeleaf_knife", () -> new SteeleafKnifeItem((new Item.Properties()).tab(FarmersDelight.CREATIVE_TAB)));
-	public static final RegistryObject<Item> LIVING = registerItem("living_knife", () -> new LivingKnifeItem((new Item.Properties()).tab(FarmersDelight.CREATIVE_TAB)));
-	public static final RegistryObject<Item> PENDORITE = registerSmithedKnife("pendorite", Util.ing(ModItems.NETHERITE_KNIFE), ingot("pendorite"), Mods.BYG);
-	public static final RegistryObject<Item> WARDEN = registerItem("warden_knife", () -> new CompatKnifeItem(Mods.DD, DelightfulItemTags.REINFORCED_ECHO_SHARD, DelightfulTiers.WARDEN, (new Item.Properties()).tab(FarmersDelight.CREATIVE_TAB), Util.ing(ModItems.NETHERITE_KNIFE), ChatFormatting.LIGHT_PURPLE));
-	public static final RegistryObject<Item> EXPERIENCE = registerItem("experience_knife", () -> new ExperienceKnifeItem((new Item.Properties()).tab(FarmersDelight.CREATIVE_TAB)));
-	public static final RegistryObject<Item> GILDED_QUARTZ = registerItem("gilded_quartz_knife", () -> new GildedQuartzKnifeItem((new Item.Properties()).tab(FarmersDelight.CREATIVE_TAB)));
-	public static final RegistryObject<Item> BLAZING = registerItem("blazing_knife", () -> new BlazingKnifeItem((new Item.Properties()).tab(FarmersDelight.CREATIVE_TAB)));
-	public static final RegistryObject<Item> LEAF = registerItem("leaf_knife", () -> new LeafKnifeItem((new Item.Properties()).tab(FarmersDelight.CREATIVE_TAB)));
-	public static final RegistryObject<Item> KIWANO = registerItem("kiwano_knife", () -> new KiwanoKnifeItem((new Item.Properties()).tab(FarmersDelight.CREATIVE_TAB)));
-	public static final RegistryObject<Item> ADAMANTITE = registerItem("adamantite_knife", () -> new AdamantiteKnifeItem((new Item.Properties()).tab(FarmersDelight.CREATIVE_TAB)));
+	public static final RegistryObject<Item> THYRIUM = registerCompatIngotKnife("thyrium", Mods.FUS);
+	public static final RegistryObject<Item> SINISITE = registerCompatIngotKnife("sinisite", Mods.FUS);
+	public static final RegistryObject<Item> ALLTHEMODIUM = registerItem("allthemodium_knife", () -> new AllthemodiumKnifeItem(props()));
+	public static final RegistryObject<Item> ENDERITE = registerItem("enderite_knife", () -> new EnderiteKnifeItem(props()));
+	public static final RegistryObject<Item> DEORUM = registerCompatIngotKnife("deorum", Mods.FA);
+	public static final RegistryObject<Item> REINFORCED_DEORUM = registerItem("reinforced_deorum_knife", () -> new ReinforcedDeorumKnifeItem(props()));
+	public static final RegistryObject<Item> DRACO_ARCANUS = registerItem("draco_arcanus_knife", () -> new DracoArcanusKnifeItem(props()));
+	public static final RegistryObject<Item> OSMIUM = registerCompatIngotKnife("osmium", Mods.MEKT);
+	public static final RegistryObject<Item> REFINED_GLOWSTONE = registerCompatIngotKnife("refined_glowstone", Mods.MEKT);
+	public static final RegistryObject<Item> REFINED_OBSIDIAN = registerCompatIngotKnife("refined_obsidian", Mods.MEKT);
+	public static final RegistryObject<Item> OBSIDIAN_INFUSED_ENDERITE = registerItem("obsidian_infused_enderite_knife", () -> new ObsdianInfusedEnderiteKnifeItem(props()));
+	public static final RegistryObject<Item> NETHERITE_OPAL = registerItem("netherite_opal_knife", () -> new NetheriteOpalKnifeItem(props()));
+	public static final RegistryObject<Item> LARGE_AMETHYST = registerItem("large_amethyst_knife", () -> new LargeAmethystKnifeItem(props()));
+	public static final RegistryObject<Item> FIERY = registerItem("fiery_knife", () -> new FieryKnifeItem(props()));
+	public static final RegistryObject<Item> IRONWOOD = registerItem("ironwood_knife", () -> new IronwoodKnifeItem(props()));
+	public static final RegistryObject<Item> KNIGHTMETAL = registerItem("knightmetal_knife", () -> new KnightmetalKnifeItem(props()));
+	public static final RegistryObject<Item> STEELEAF = registerItem("steeleaf_knife", () -> new SteeleafKnifeItem(props()));
+	public static final RegistryObject<Item> LIVING = registerItem("living_knife", () -> new LivingKnifeItem(props()));
+	public static final RegistryObject<Item> PENDORITE = registerItem("pendorite_knife", () -> new PendoriteKnifeItem(props()));
+	public static final RegistryObject<Item> WARDEN = registerItem("warden_knife", () -> new WardenKnifeItem(props()));
+	public static final RegistryObject<Item> EXPERIENCE = registerItem("experience_knife", () -> new ExperienceKnifeItem(props()));
+	public static final RegistryObject<Item> GILDED_QUARTZ = registerItem("gilded_quartz_knife", () -> new GildedQuartzKnifeItem(props()));
+	public static final RegistryObject<Item> BLAZING = registerItem("blazing_knife", () -> new BlazingKnifeItem(props()));
+	public static final RegistryObject<Item> LEAF = registerItem("leaf_knife", () -> new LeafKnifeItem(props()));
+	public static final RegistryObject<Item> KIWANO = registerItem("kiwano_knife", () -> new KiwanoKnifeItem(props()));
+	public static final RegistryObject<Item> ADAMANTITE = registerItem("adamantite_knife", () -> new AdamantiteKnifeItem(props()));
 	public static final RegistryObject<Item> CRYSTALLINE = registerCompatKnife("crystalline", Mods.EP, DelightfulItemTags.VOID_CRYSTAL_BLOCK);
-	public static final RegistryObject<Item> STELLIUM = registerItem("stellium_knife", () -> new StelliumKnifeItem((new Item.Properties()).tab(FarmersDelight.CREATIVE_TAB)));
+	public static final RegistryObject<Item> STELLIUM = registerItem("stellium_knife", () -> new StelliumKnifeItem(props()));
+	public static final RegistryObject<Item> SOUL_STEEL = registerItem("soul_steel_knife", () -> new SoulSteelKnifeItem(props()));
+	public static final RegistryObject<Item> ROSE_GOLD = registerItem("rose_gold_knife", () -> new RoseGoldKnifeItem(props()));
+	public static final RegistryObject<Item> GILDED_NETHERITE = registerItem("gilded_netherite_knife", () -> new GildedNetheriteKnifeItem(props()));
+	public static final RegistryObject<Item> NECRONIUM = registerItem("necronium_knife", () -> new NecroniumKnifeItem(props()));
+	public static final RegistryObject<Item> CLOGGRUM = registerCompatIngotKnife("cloggrum", Mods.UG);
+	public static final RegistryObject<Item> FROSTSTEEL = registerItem("froststeel_knife", () -> new FroststeelKnifeItem(props()));
+	public static final RegistryObject<Item> UTHERIUM = registerItem("utherium_knife", () -> new UtheriumKnifeItem(props()));
+	public static final RegistryObject<Item> FORGOTTEN = registerItem("forgotten_knife", () -> new ForgottenKnifeItem(props()));
+	public static final RegistryObject<Item> MANASTEEL = registerItem("manasteel_knife",() -> new ManasteelKnifeItem(
+		props(),
+		DelightfulItems.ingot("manasteel"),
+		Mods.loaded(Mods.BTA) ? BotaniaCompat.manasteel().get() : DelightfulTiers.STEEL
+	));
+	public static final RegistryObject<Item> ELEMENTIUM = registerItem("elementium_knife", () -> new ElementiumKnifeItem(
+		props(),
+		DelightfulItems.ingot("elementium"),
+		Mods.loaded(Mods.BTA) ? BotaniaCompat.elementium().get() : DelightfulTiers.STEEL
+	));
+	public static final RegistryObject<Item> TERRA = registerItem("terra_knife", () -> new TerraKnifeItem(
+		props(),
+		DelightfulItems.ingot("terrasteel"),
+		Mods.loaded(Mods.BTA) ? BotaniaCompat.terrasteel().get() : DelightfulTiers.STEEL,
+		ChatFormatting.YELLOW
+	));
+	public static final RegistryObject<Item> ALF = registerItem("alf_knife", () -> new AlfKnifeItem(
+		props()
+	));
+
+	private static Item.Properties props() {
+		return (new Item.Properties()).tab(FarmersDelight.CREATIVE_TAB);
+	}
 
 	// Registers a knife to Farmer's Delight tab, requiring modid
-	public static RegistryObject<Item> registerCompatKnife(String name, String modid, TagKey<Item> tag, Component... tool) {
-		if (tool.length > 0) {
-			return registerItem(name + "_knife", () -> new CompatKnifeItem(modid, tag, DelightfulTiers.get(name), (new Item.Properties()).tab(FarmersDelight.CREATIVE_TAB), tool[0], null));
-		} else {
-			return registerItem(name + "_knife", () -> new CompatKnifeItem(modid, tag, DelightfulTiers.get(name), (new Item.Properties()).tab(FarmersDelight.CREATIVE_TAB), null));
-		}
+	public static RegistryObject<Item> registerCompatKnife(String name, String modid, TagKey<Item> tag) {
+		return registerItem(name + "_knife", () -> new CompatKnifeItem(modid, tag, DelightfulTiers.get(name), (new Item.Properties()).tab(FarmersDelight.CREATIVE_TAB)));
+	}
+
+	// Registers a knife to Farmer's Delight tab, requiring non-empty ingot tag
+	public static RegistryObject<Item> registerCompatIngotKnife(String name, String modid) {
+		return registerItem(name + "_knife", () -> new CompatKnifeItem(modid, ingot(name), DelightfulTiers.get(name), (new Item.Properties()).tab(FarmersDelight.CREATIVE_TAB)));
 	}
 
 	// Registers a knife to Farmer's Delight tab, requiring non-empty ingot tag
 	public static RegistryObject<Item> registerIngotKnife(String name) {
-		return registerItem(name + "_knife", () -> new DelightfulKnifeItem(ingot(name), DelightfulTiers.get(name), (new Item.Properties()).tab(FarmersDelight.CREATIVE_TAB), null));
+		return registerItem(name + "_knife", () -> new DelightfulKnifeItem(ingot(name), DelightfulTiers.get(name), (new Item.Properties()).tab(FarmersDelight.CREATIVE_TAB)));
 	}
 
 	// Registers a knife to Farmer's Delight tab, requiring non-empty gem tag
 	public static RegistryObject<Item> registerGemKnife(String name) {
-		return registerItem(name + "_knife", () -> new DelightfulKnifeItem(gem(name), DelightfulTiers.get(name), (new Item.Properties()).tab(FarmersDelight.CREATIVE_TAB), null));
-	}
-
-	// Registers a knife to Farmer's Delight tab, requiring non-empty ingot tag
-	public static RegistryObject<Item> registerSmithedKnife(String name, Supplier<Ingredient> base, TagKey<Item> addition, String... modid) {
-		if (modid.length > 0) {
-			return registerItem(name + "_knife", () -> new CompatKnifeItem(modid[0], addition, DelightfulTiers.get(name), (new Item.Properties()).tab(FarmersDelight.CREATIVE_TAB), base));
-		} else {
-			return registerItem(name + "_knife", () -> new DelightfulKnifeItem(addition, DelightfulTiers.get(name), (new Item.Properties()).tab(FarmersDelight.CREATIVE_TAB), base));
-		}
+		return registerItem(name + "_knife", () -> new DelightfulKnifeItem(gem(name), DelightfulTiers.get(name), (new Item.Properties()).tab(FarmersDelight.CREATIVE_TAB)));
 	}
 
 	// Registers a knife to Farmer's Delight tab
 	public static RegistryObject<Item> registerKnife(String name, TagKey<Item> tag) {
-		return registerItem(name + "_knife", () -> new DelightfulKnifeItem(tag, DelightfulTiers.get(name), (new Item.Properties()).tab(FarmersDelight.CREATIVE_TAB), null));
+		return registerItem(name + "_knife", () -> new DelightfulKnifeItem(tag, DelightfulTiers.get(name), (new Item.Properties()).tab(FarmersDelight.CREATIVE_TAB)));
 	}
 
 	public static void create() {}