+++ src/main/java/net/brnbrd/delightful/compat/ArsNouveauCompat.java
@@ -7,13 +7,10 @@ import java.util.function.Supplier;
 public class ArsNouveauCompat {
 
 	public static final String pie = "source_berry_pie";
-	public static final String slice = "source_berry_pie_slice";
 
 	public static Supplier<FoodProperties> getPieSlice() {
-		if (Mods.loaded(Mods.AN)) {
-			return ArsNouveauSuppliers.PIE_SLICE;
-		} else {
-			return () -> FoodValues.PIE_SLICE;
-		}
+		return (Mods.loaded(Mods.AN)) ?
+			ArsNouveauSuppliers.PIE_SLICE :
+			() -> FoodValues.PIE_SLICE;
 	}
 }