+++ src/main/java/net/brnbrd/delightful/compat/Mods.java
@@ -10,6 +10,7 @@ import java.util.function.Supplier;
 public class Mods {
 	public static final String AN = "ars_nouveau";
 	public static final String BOP = "biomesoplenty";
+	public static final String BTA = "botania";
 	public static final String BYG = "byg";
 	public static final String DD = "deeperdarker";
 	public static final String ECO = "ecologics";
@@ -28,6 +29,7 @@ public class Mods {
 	public static final String RL = "rottenleather";
 	public static final String SO = "simpleores";
 	public static final String TF = "twilightforest";
+	public static final String UG = "undergarden";
 	public static final String WB = "wildberries";
 
 	public static boolean loaded(@NotNull String modid) {
@@ -41,6 +43,15 @@ public class Mods {
 			}
 		}
 		return true;
+	}
+
+	public static boolean orLoaded(@NotNull String... modids) {
+		for (String mod : modids) {
+			if (loaded(mod)) {
+				return true;
+			}
+		}
+		return false;
 	}
 
 	public static Supplier<MobEffect> getVitality() {